requirements.txt
# Core dependencies
openai>=1.0.0
playwright>=1.40.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and automation
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
pathlib2>=2.3.7
hashlib2>=1.0.1

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: for enhanced logging
rich>=13.0.0
loguru>=0.7.0